:memo: :: Change Type(Long-id)

+++ presentation/src/main/java/com/chobo/presentation/view/component/combinationView/CombinationViewNavigation.kt
@@ -14,7 +14,7 @@ fun NavController.navigationToCombinationView() {
 
 fun NavGraphBuilder.combinationView(
     topDestination: MutableState<MindWayNavBarItemType>,
-    navigateToDetailEvent: () -> Unit,
+    navigateToDetailEvent: (Long) -> Unit,
     navigateToGoalReading: () -> Unit,
     navigateToBookAddBook: () -> Unit,
     navigateToIntro: () -> Unit,
